daa/requirements.txt
Django==1.8
django-admin-tools==0.8.1
django-debug-toolbar==1.9.1
django-docutils==0.4.0
django-filter==1.1.0
django-filters==0.2.1
django-markdown-deux==1.0.5
django-markwhat==1.6.0
django-mptt==0.8.7
django-reversion==2.0.12
django-suit==0.2.5
django-tables2==1.16.0
django-utils==0.0.2
docutils==0.14
encoding==1.0.0
lxml==4.1.1
markdown2==2.3.5
markup==0.2
MySQL-python==1.2.5
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
simplejson==3.2.0
six==1.11.0
sqlparse==0.2.4
